FEAT: Add routing to allow navigation between pages
With 3 pages we need 3 methods in 'run.py' - one method to display each page.
In the html code we inject a 'url_for()' jinja method into the href for each link.
We pass the appropriate method name into each url_for().
This gives us a link to each page in the navigation section.